requirements.txt
# Python AI Service Dependencies
# FastAPI + Uvicorn
fastapi>=0.104.0,<1.0
uvicorn[standard]>=0.24.0,<1.0
pydantic>=2.5.0,<3.0
pydantic-settings>=2.1.0,<3.0

# OpenAI Embedding + LLM
openai>=1.6.0,<2.0

# Anthropic Claude LLM
anthropic>=0.39.0,<1.0

# Milvus Vector DB
pymilvus>=2.3.4,<3.0

# LangChain (text splitting + document loading)
langchain-text-splitters>=0.0.1,<1.0
langchain-community>=0.0.10,<1.0
pypdf>=3.17.0,<5.0

# Document parsing
python-docx>=1.1.0,<2.0
beautifulsoup4>=4.12.0,<5.0
requests>=2.31.0,<3.0

# Retry logic
tenacity>=8.2.0,<10.0

# Multipart form support
python-multipart>=0.0.6,<1.0
